feat: 자동차 이름 및 시도 횟수 입력 기능 구현
- 사용자로부터 경주할 자동차의 이름을 입력받는 기능 추가
- 입력받은 이름을 기반으로 자동차 객체 생성 및 리스트에 저장
- 사용자로부터 시도할 횟수를 입력받는 기능 추가
- 입력된 시도 횟수의 유효성 검사 및 예외 처리 추가
- 컨트롤러에서 입력받은 데이터를 처리하여 레이스 시작 기능 구현

+++ src/main/java/racingcar/model/Race.java
@@ -1,15 +1,19 @@
 package racingcar.model;
 
+
+import java.math.BigInteger;
 import java.util.List;
 
+
 public class Race {
-    private List<Car> cars;
-    private int attempts;
+   private List<Car> cars;
+   private BigInteger attempts;
 
 
-    public Race(List<Car> cars, int attempts) {
-        this.cars = cars;
-        this.attempts = attempts;
-    }
+   public Race(List<Car> cars, BigInteger attempts) {
+       this.cars = cars;
+       this.attempts = attempts;
+      
+   }
 
 }
